+++ scripts/generate-map-geometry/lib/utils.ts
@@ -1,6 +1,9 @@
 import { lineWidth } from "@/components/map/renderer/utils";
 import { flexi } from "@/scripts/generate-map-geometry/lib/dimensions/flexi-length";
 import { SegmentInstruction } from "@/scripts/generate-map-geometry/lib/segment-instructions";
+import { FlexiPoint } from "@/scripts/generate-map-geometry/lib/dimensions/flexi-point";
+import { LineBuilder } from "@/scripts/generate-map-geometry/lib/line-builder";
+import { SegmentBuilder } from "@/scripts/generate-map-geometry/lib/segment-builder";
 
 export const interchangeEdgeOffset = flexi(lineWidth / 2);
 export const interchangeInnerOffset = flexi(1);
@@ -21,4 +24,22 @@ export function invert(
     .reverse();
 
   return reversed;
+}
+
+export function getNodePosition(line: LineBuilder, nodeId: number): FlexiPoint {
+  const locatedNode = line.build().nodes.find((n) => n.nodeId === nodeId);
+  if (!locatedNode) {
+    throw new Error(`Node ${nodeId} not found`);
+  }
+  return locatedNode.point;
+}
+
+export function getEndPoint(
+  startPoint: FlexiPoint,
+  startAngle: number,
+  instructions: SegmentInstruction[],
+): FlexiPoint {
+  return new SegmentBuilder(startPoint, startAngle)
+    .process(instructions)
+    .build().endPoint;
 }
