+++ scripts/generate-map-geometry/ptv/interchanges.ts
@@ -2,8 +2,6 @@ import { InterchangeBlueprint } from "@/scripts/generate-map-geometry/lib/interc
 import * as station from "@/shared/station-ids";
 import * as node from "@/shared/map-node-ids";
 
-// TODO: [DS] Split each interchange into it's own file.
-
 export const flindersStreet = InterchangeBlueprint.simple(
   station.FLINDERS_STREET,
   [
