+++ scripts/generate-map-geometry/ptv/lines/northern.ts
@@ -68,6 +68,7 @@ export const northern = new LineBuilder(
     l
       // TODO: One day we might show the Flemington Racecourse line, and if so,
       // we'll want to add the node for Newmarket here.
+
       .to(node.BROADMEADOWS, [
         straight(newmarketStraight),
         curve(newmarketCurveCraigieburn, 45),
