+++ scripts/generate-map-geometry/ptv/segments/southern-cross-to-north-melbourne.ts
@@ -13,7 +13,7 @@ import {
   SegmentInstruction,
   straight,
 } from "@/scripts/generate-map-geometry/lib/segment-instructions";
-import { getEndPoint } from "@/scripts/generate-map-geometry/lib/measure";
+import { getEndPoint } from "@/scripts/generate-map-geometry/lib/utils";
 
 const innerRadius = flexi(15);
 const loopDirectRadius = flexi(20);
